+++ src/panorama/utils/colors.ts
@@ -1,0 +1,7 @@
+export const colors = {
+  orange: '#ff4500',
+  gold: '#ffd700',
+  purple: '#a029af',
+  blue: '#1e90ff ',
+  grey: '#d3d3d3',
+};
