+++ src/panorama/utils/utils.ts
