+++ src/panorama/webpack.dev.js
@@ -29,6 +29,9 @@ module.exports = {
 
   resolve: {
     extensions: ['.ts', '.tsx', '.js', '.jsx', '...'],
+    alias: {
+      '@utils': path.resolve(__dirname, 'utils'),
+    },
     symlinks: false,
   },
